magic proxies for all token types
