+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "magic_bridge/ic/src/DIP20"]
-	path = magic_bridge/ic/src/DIP20
+[submodule "magic_bridge/ic/DIP20"]
+	path = magic_bridge/ic/DIP20
 	url = git@github.com:Psychedelic/DIP20.git
