+++ magic_bridge/ic/DIP20/motoko/demo.sh
