+++ magic_bridge/ic/DIP20/motoko/dfx.json
