+++ magic_bridge/ic/DIP20/motoko/src/cap/Root.mo
