+++ magic_bridge/ic/DIP20/motoko/src/token.mo
