+++ magic_bridge/ic/DIP20/rust/Cargo.toml
