+++ magic_bridge/ic/DIP20/rust/deploy.sh
