+++ magic_bridge/ic/DIP20/rust/dfx.json
