+++ magic_bridge/ic/DIP20/rust/src/main.rs
