+++ magic_bridge/ic/DIP20/rust/token.did
