+++ magic_bridge/ic/DIP20/scripts/token.sh
