+++ magic_bridge/ic/factory.sh
