--- magic_bridge/ic/src/factory/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "factory"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-ic-cdk-macros = "0.3"
-candid = "0.7.4"
-serde = "1.0"
-serde_cbor = "0.11.2"
-serde_bytes = "0.11.5"
-ic-kit = "0.4.3"
-ic-cdk = "0.3.1"
-assert-panic = "1.0.1"
-
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-async-std = { version="1.10.0", features = ["attributes"] }
