--- magic_bridge/ic/src/factory/dfx.json
@@ -1,22 +1,0 @@
-{
-    "canisters": {
-        "factory": {
-            "build": "cargo build --target wasm32-unknown-unknown --release",
-            "candid": "factory.did",
-            "wasm": "target/wasm32-unknown-unknown/release/factory.wasm",
-            "type": "custom"
-        }
-    },
-    "defaults": {
-      "build": {
-        "packtool": ""
-      }
-    },
-    "networks": {
-      "local": {
-        "bind": "127.0.0.1:8000",
-        "type": "ephemeral"
-      }
-    },
-    "version": 1
-  }
