--- magic_bridge/ic/src/factory/factory.did
@@ -1,32 +1,0 @@
-type FactoryError = variant {
-  CanisterStatusNotAvailableError;
-  CreateCanisterError;
-  EncodeError;
-  CodeAlreadyInstalled;
-  InstallCodeError;
-};
-
-type TokenType = variant {
-  DIP20;
-  DIP721;
-  DIP1155;
-};
-
-type Result = variant { Ok : principal; Err : FactoryError };
-service : (principal) -> {
-  create : (
-      text,
-      text,
-      text,
-      nat8,
-      nat,
-      principal,
-      vec principal,
-      nat64,
-      nat,
-      principal,
-      principal,
-      TokenType,
-    ) -> (Result);
-  owner : () -> (principal) query;
-}
