+++ magic_bridge/ic/src/magic_bridge/magic_bridge.did
@@ -1,3 +1,32 @@
-service : {
-    "greet": (text) -> (text) query;
+type FactoryError = variant {
+  CanisterStatusNotAvailableError;
+  CreateCanisterError;
+  EncodeError;
+  CodeAlreadyInstalled;
+  InstallCodeError;
+};
+
+type TokenType = variant {
+  DIP20;
+  DIP721;
+  DIP1155;
+};
+
+type Result = variant { Ok : principal; Err : FactoryError };
+service : (principal) -> {
+  create : (
+      text,
+      text,
+      text,
+      nat8,
+      nat,
+      principal,
+      vec principal,
+      nat64,
+      nat,
+      principal,
+      principal,
+      TokenType,
+    ) -> (Result);
+  owner : () -> (principal) query;
 }
