+++ magic_bridge/ic/src/magic_bridge/proxy.rs
@@ -1,0 +1,28 @@
+use ic_kit::{
+    candid::{Nat, Principal},
+    ic, RejectionCode,
+};
+
+pub struct Dip20Proxy {}
+
+impl Dip20Proxy {
+    pub async fn create(
+    ) {
+    }
+}
+
+pub struct Dip721Proxy {}
+
+impl Dip721Proxy {
+    pub async fn create(
+    ) {
+    }
+}
+
+pub struct Dip1155Proxy {}
+
+impl Dip1155Proxy {
+    pub async fn create(
+    ) {
+    }
+}
