+++ magic_bridge/ic/src/magic_bridge/wasm/dip20/token.wasm
